Mejoras: Gestión de usuarios administradores y sincronización de ejercicios
- Se agregó filtrado de roles para usuarios administradores en UserController e IndexUserRequest.

- Se implementó un endpoint para las próximas membresías que ahora acepta un parámetro de días enteros.

- Se creó una nueva pantalla de ejercicios para gestionar el catálogo de ejercicios con función de sincronización.

- Se introdujo el componente SearchBar para mejorar la funcionalidad de búsqueda en todas las pantallas.

- Se actualizó AdminDashboardScreen para incluir acceso rápido a ejercicios y usuarios administradores.

- Se mejoró la interfaz de usuario de gestión de usuarios con controles segmentados para usuarios y administradores.

- Se agregó funcionalidad de sincronización para ejercicios mediante integración de API.

- Se actualizaron las rutas y la navegación para la nueva pestaña de ejercicios en el diseño de administración.

+++ backend/app/Http/Controllers/Api/V1/Admin/MembershipController.php
@@ -29,7 +29,7 @@ class MembershipController extends Controller
     public function upcoming(MembershipIndexRequest $request): JsonResponse
     {
         $filters = $request->validated();
-        $days = $filters['days'] ?? 30;
+        $days = (int) ($filters['days'] ?? 30);
 
         $memberships = Membership::query()
             ->with('user')
